requirement.txt
Flask==2.1.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
bcrypt==3.2.2
